Crate cu29_rendercfg

Modules

config 🔒

This module defines the configuration of the copper runtime. The configuration is a directed graph where nodes are tasks and edges are connections between tasks. The configuration is serialized in the RON format. The configuration is used to generate the runtime code at compile time.

Structs

Args 🔒

CuError

Common copper Error type.

Enums

UnifiedLogType

Defines the types of what can be logged in the unified logger.

Traits

CopperListTuple

A CopperListTuple needs to be encodable, decodable and fixed size in memory.

WriteStream

Defines a basic write, append only stream trait to be able to log or send serializable objects.

Functions

main 🔒

Render the configuration file to a dot file then convert it to an SVG and optionally opens it with inkscape.

Type Aliases

CuResult