cu29_rendercfg::config

Struct Node

pub struct Node {
    id: String,
    type_: Option<String>,
    config: Option<ComponentConfig>,
    resources: Option<HashMap<String, String>>,
    missions: Option<Vec<String>>,
    background: Option<bool>,
    run_in_sim: Option<bool>,
    logging: Option<NodeLogging>,
    flavor: Flavor,
}

A node in the configuration graph. A node represents a Task in the system Graph.

Fields
id: String

Unique node identifier.

type_: Option<String>

Task rust struct underlying type, e.g. โmymodule::Sensorโ, etc.

config: Option<ComponentConfig>

Config passed to the task.

resources: Option<HashMap<String, String>>

Resources requested by the task.

missions: Option<Vec<String>>

Missions for which this task is run.

background: Option<bool>

Run this task in the background: ie. Will be set to run on a background thread and until it is finished CuTask::process will return None.

run_in_sim: Option<bool>

Option to include/exclude stubbing for simulation. By default, sources and sinks are replaces (stubbed) by the runtime to avoid trying to compile hardware specific code for sensing or actuation. In some cases, for example a sink or source used as a middleware bridge, you might want to run the real code even in simulation. This option allows to control this behavior. Note: Normal tasks will be run in sim and this parameter ignored.

logging: Option<NodeLogging>

Config passed to the task.

flavor: Flavor

Node role in the runtime graph (normal task or bridge endpoint).

Implementations

impl Node

pub fn new(id: &str, ptype: &str) -> Self

pub fn new_with_flavor(id: &str, ptype: &str, flavor: Flavor) -> Self

pub fn get_id(&self) -> String

pub fn get_type(&self) -> &str

pub fn set_type(self, name: Option<String>) -> Self

pub fn set_resources<I>(&mut self, resources: Option<I>)
where I: IntoIterator<Item = (String, String)>,

pub fn is_background(&self) -> bool

pub fn get_instance_config(&self) -> Option<&ComponentConfig>

pub fn get_resources(&self) -> Option<&HashMap<String, String>>

pub fn is_run_in_sim(&self) -> bool

By default, assume a source or a sink is not run in sim. Normal tasks will be run in sim and this parameter ignored.

pub fn is_logging_enabled(&self) -> bool

pub fn get_param<T>(&self, key: &str) -> Result<Option<T>, ConfigError>
where T: for<'a> TryFrom<&'a Value, Error = ConfigError>,

pub fn set_param<T: Into<Value>>(&mut self, key: &str, value: T)

pub fn get_flavor(&self) -> Flavor

Returns whether this node is treated as a normal task or as a bridge.

pub fn set_flavor(&mut self, flavor: Flavor)

Overrides the node flavor; primarily used when injecting bridge nodes.