Crate cu29_rendercfg

Modules

config 🔒

This module defines the configuration of the copper runtime. The configuration is a directed graph where nodes are tasks and edges are connections between tasks. The configuration is serialized in the RON format. The configuration is used to generate the runtime code at compile time.

Structs

Args 🔒

CuCompactString

CuError

Common copper Error type.

Enums

UnifiedLogType

Defines the types of what can be logged in the unified logger.

Constants

COMPACT_STRING_CAPACITY

Traits

CopperListTuple

A CopperListTuple needs to be encodable, decodable and fixed size in memory.

CuMsgMetadataTrait

Key metadata piece attached to every message in Copper.

ErasedCuStampedData

A generic trait to expose the generated CuStampedDataSet from the task graph.

ErasedCuStampedDataSet

Trait to get a vector of type-erased CuStampedDataSet This is used for generic serialization of the copperlists

MatchingTasks

Trait to trace back from the CopperList the origin of the messages

Metadata

Represent the minimum set of traits to be usable as Metadata in Copper.

WriteStream

Defines a basic write, append only stream trait to be able to log or send serializable objects.

Functions

main 🔒

Render the configuration file to a dot file then convert it to an SVG and optionally opens it with inkscape.

Type Aliases

CuResult