Trait cu29_rendercfg::WriteStream

pub trait WriteStream<E>:
    Sync
    + Send
    + Debug
where
    E: Encode,
{
    // Required method
    fn log(&mut self, obj: &E) -> Result<(), CuError>;

    // Provided method
    fn flush(&mut self) -> Result<(), CuError> { ... }
}

Defines a basic write, append only stream trait to be able to log or send serializable objects.

Required Methods

fn log(&mut self, obj: &E) -> Result<(), CuError>

Provided Methods

fn flush(&mut self) -> Result<(), CuError>